refactor: Move bookmark store to @/stores/ sub-dir
